+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Breaking changes:
+  - The filter now returns HTTP 401 'Unauthorized' instead of HTTP 403 'Forbidden' when no API Key or an invalid API Key is provided.
+
 ## [2.0.0] - 2023-01-31
 - Breaking changes: This version requires Java 17 and Spring Boot 3.0.2+
   - Due to the switch from Java EE to Jakarta EE, this version of the library is incompatible with Spring Boot versions prior to 3.0.0.